Refactor tameable entities
- Abstract Ownable entities in a separate interface
- Make TraderLlama extend Llama
- Make Camel Sittable
- Rename Keys.Tamer to Keys.OWNER
