+++ src/main/java/org/spongepowered/api/entity/living/animal/horse/HorseLike.java
@@ -26,25 +26,15 @@ package org.spongepowered.api.entity.living.animal.horse;
 
 import org.spongepowered.api.data.Keys;
 import org.spongepowered.api.data.value.Value;
+import org.spongepowered.api.entity.Ownable;
 import org.spongepowered.api.entity.living.animal.Animal;
 import org.spongepowered.api.item.inventory.Carrier;
 
-import java.util.Optional;
-import java.util.UUID;
 
 /**
  * An abstract representation of a Horse.
  */
-public interface HorseLike extends Animal, Carrier {
-
-    /**
-     * {@link Keys#IS_SADDLED}
-     *
-     * @return Whether the horse is saddled
-     */
-    default Value.Mutable<Boolean> saddled() {
-        return this.requireValue(Keys.IS_SADDLED).asMutable();
-    }
+public interface HorseLike extends Animal, Ownable, Carrier {
 
     /**
      * {@link Keys#IS_TAMED}
@@ -56,12 +46,12 @@ public interface HorseLike extends Animal, Carrier {
     }
 
     /**
-     * {@link Keys#TAMER}
+     * {@link Keys#IS_SADDLED}
      *
-     * @return The tamer of the horse
+     * @return Whether the horse is saddled
      */
-    default Optional<Value.Mutable<UUID>> tamer() {
-        return this.getValue(Keys.TAMER).map(Value::asMutable);
+    default Value.Mutable<Boolean> saddled() {
+        return this.requireValue(Keys.IS_SADDLED).asMutable();
     }
 
 }
