+++ src/main/java/org/spongepowered/api/entity/living/animal/horse/llama/Llama.java
@@ -28,14 +28,13 @@ import org.spongepowered.api.data.Keys;
 import org.spongepowered.api.data.type.LlamaType;
 import org.spongepowered.api.data.value.Value;
 import org.spongepowered.api.entity.living.animal.horse.PackHorse;
-import org.spongepowered.api.entity.living.animal.horse.TameableHorse;
 
 /**
  * Represents a Llama. Llamas are unique in that
  * they can be ridden by players, but not controlled by players. Likewise, they
  * have colors and styles and can have storage "strength".
  */
-public interface Llama extends LlamaLike, TameableHorse, PackHorse {
+public interface Llama extends LlamaLike, PackHorse {
 
     /**
      * {@link Keys#LLAMA_TYPE}
