+++ src/main/java/org/spongepowered/api/entity/living/animal/horse/llama/TraderLlama.java
@@ -28,7 +28,10 @@ import org.spongepowered.api.data.Keys;
 import org.spongepowered.api.data.value.Value;
 import org.spongepowered.api.util.Ticks;
 
-public interface TraderLlama extends LlamaLike {
+/**
+ * Represents a TraderLlama.
+ */
+public interface TraderLlama extends Llama {
 
     /**
      * {@link Keys#DESPAWN_DELAY}
